+++ scripts/constants.js
@@ -1,0 +1,17 @@
+// for storage
+export const GAPI_LOADED = "gapi_loaded",
+  ALL_SUBS = "all_subs",
+  LAST_SYNCED = "last_synced",
+  REDUNDANT_EMAILS = "redundant_emails",
+  START = "start";
+
+// message types
+
+export const OPEN_NEW_TAB = "open_new_tab",
+  SYNC = "sync",
+  UPDATED_SUBSCRIBERS = "updated_subscribers",
+  RESET = "reset";
+
+// gapi
+export const DISCOVERY_URL = "https://gmail.googleapis.com/$discovery/rest?version=v1",
+  GAPI_CLIENT_URL = "https://apis.google.com/js/client.js";